convert CURLFTPMETHOD_* to long constants

+++ src/tool_cfgable.h
@@ -227,7 +227,7 @@ struct OperationConfig {
   HttpReq httpreq;
   long proxyver;             /* set to CURLPROXY_HTTP* define */
   long ftp_ssl_ccc_mode;
-  int ftp_filemethod;
+  long ftp_filemethod;
   enum {
     CLOBBER_DEFAULT, /* Provides compatibility with previous versions of curl,
                         by using the default behavior for -o, -O, and -J.
